app/requirements.txt
annotated-types==0.5.0
anyio==3.7.1
click==8.1.5
exceptiongroup==1.1.2
fastapi==0.100.0
h11==0.14.0
idna==3.4
pydantic==2.0.2
pydantic-core==2.1.2
python-multipart==0.0.6
sniffio==1.3.0
starlette==0.27.0
typing-extensions==4.7.1
uvicorn==0.22.0
Pillow
pandas==1.5.3
opencv-python
numpy==1.23.5
pyocr
openpyxl
img2table
Levenshtein==0.21.1
pytesseract==0.3.10
